Cookies are working!
The script to validate a cookie for purposes of validating the admin
flag is loadNav.js, which should now be in the script list at the
bottom of every HTML page.
loginValidate.py validates the login and sets the cookies now. We
should perhaps determine what needs to be set for cookies and what we
don’t set for cookies.